Added generic view for JSONP, modified GET method.
Addresses issue #9.

The GET method for fetching a study from OTOL API now passes a dict (as
required) to our generic JSON/P views. These now check for a
FULL_RESPONSE key in the dict; if found, the view renders this value and
ignores the surrounding dict.

The added 'generic.jsonp' view lets us support cross-domain API
requests. This should be triggered by the presence of a 'callback' or
'jsoncallback' argument in the request (see ```default.py``` controller
for an example).