Reading Github app info from private config
This allows us to keep our higher rate limit with the Github APIs when using anonymous requests, eg, for an anonymous visitor viewing a study.  I'm using the registered client ID and client secret for the OpenTree curation app, but we could register the OTOL API separately if needed to support heavy traffic.