Properly create trees on Github and go don't indent our JSON for now
Tree paths should not start with a leading forward slash. The Github
API will allow you to create them successfully, but they create a corrupt
tree which prevents the repo from being clonable.

Deleting the branch containing the offending tree caused the repo to be
clonable again.

Previously we were indenting JSON 4 spaces, but all current JSON in treenexus
has indent=0. With this change, we get meaningful diffs of the changed JSON.
For example:

    https://github.com/OpenTreeOfLife/treenexus/compare/leto_study_9?expand=1