translating the NexSON on the fly in v1/study GET
Note the new prerequisite and the new configuration
in the README. This set of changes supports a
new output_nexml2json parameter to the GET of a
study. This argument can take the values 0.0.0 and
1.0.0 at this point.

Caveats:
    The translation is very slow for large
studies. The translation between 0.0.0 and 1.0.0
currently goes through NeXML (using this current
version peyotl).
    The PUT and POST verbs are not tolerant
of different forms of NexSON yet.