obscure bug fix: publication_DOI can be null
This cropped up when testing the previous commit - it probably
does not affect any real client code, because clients would
omit publication_DOI rather than sending in a null value.
A test script sent in a null value, without the check in this commit
that lead to an internal server error.