Returning 400 if the resource_id to push is bogus
This exception deals with bogus requests nicely (failure
to put the resource_id in the right spot in the URL).
If we have some sort of deep problem that causes the
phylesystem to be not pushing (or being able to create git_actions
at all), then these will NOT be logged. So we won't be able
to detect them with our pushes_succeeding check.  However,
such a state would probably not be responding to GETs, so
we'd probably notice in other ways...