Rough cut of patching collection URLs from v1.
This is crude, but it can be tested on devapi. If it runs properly,
it should have no effect on existing features but will respond to
requests (all HTTP methods) to /v2/collections/... and /v2/collection/{ID}
In all cases, the browser should see tiny text messages as the response
payload.