Minor refactoring and logging of an s for skipped
Sending the check for read-only through a function
in the opetreetesting.py so that an 's' can be shown
for every test skipped.

In the python standard testing each test will show an s for
skipped, a . for passed, and a F for failed.

With this tweak, the user will see s for skipped by (because
of a clean exit) the run_tests.sh will still show a . afterward.
This is a bit sub-optimal, but at least lets the user see that
some of the tests had elements that were skipped.