Treat incoming Markdown as a named argument.
This is expected by web2py.

N.B. It's not clear to me why the old version worked at all! There are
no significant code changes (server- or client-side) since it was known
to be working, so I'm guessing it's some kind of change in routing
behavior, or something to do with redirects to HTTPS...? Very strange.