Reduce dependence on conf files.
These 2 tests used to get the parentsha for a commit
from the test.conf file. That worked and isolated GET
failure from these PUT tests. But it is tedious to keep
the parentsha in sync with the server against which you
are running the tests. Given that we really need the
GET to be working to use the system at all, I've gone
ahead and made these tests get the parentsha from the
response of the GET rather than the conf file.